Refactor training script to include testing samples
This commit refactors the training script in order to include testing samples in the training process. The `create_and_train` function now takes an additional argument `num_samples_test` to specify the number of testing samples. The testing samples are generated using the `SyntheticBinaryValued` dataset class and are loaded into a separate `DataLoader` called `dataloader_test`. The loss and accuracy of the testing samples are also calculated and logged in the `logs` dataframe.